MDL-85836 assign: refactor submit code, so the back-end is reusable
Previously, the code in mod_assign that handled submission had a
single method which took the submitted form data and verified all
the normal logic for 'can the this user submit this now' with the
acutal processing to update the state of the submission in the database
and send notifications, update completion, etc.

This was problematic, for example if you were trying to write a custom
scheduled_task which automatically turned draft submissions into final
submissions in certain circumstances. Hence this re-factor.