MDL-85836 assign: document a useless submit_for_grading argument
This never worked, so was potentially misleading, so I have cleaned this up.

However, I now notice that this pattern is common in other methods in this
class. I am not going to attempt a full clean-up now. I will leave that
for the mod_assign maintainer. However, I do want to clean up this one
method before I change it in the next commit.