MDL-86066 reportbuilder: schedule schema update to support class types.
In order to support upcoming changes that allow for multiple types of
report schedule to be configured, update the table/persistent schema to
be more generic.

Audience data is retained as a first-class property, in order to more
easily and consistently support schedules that make use of audience data
and API, although the field itself is now nullable.