MDL-73602 questions: Prevent restore of qtype_random questions
This completely removes the step of creating a question with qtype
'random' during the restore process.

Previously, we would create the question bank entry, question version
and question record, then rely on the activity using it to convert it to
a question set reference and delete it. If the activity was never
restored, the question remained in the database.

Now, we store the random question data in a temporary record. The
activity still converts this to a set reference, and now directly to the
new filter condition format. The temporary record will automatically be
cleaned up after the restore.

If the restore process attempts to create a question record with qtype
'random', this will now throw a coding exception as this must never
happen.