MDL-73602 quiz: Restore random tag filters to new format
As we are now creating set references using the modern filtercondition
format, this ensures legacy tag filters are also converted directly to
the new format.

We no longer need the tag name as part of this, only the ID, so we have
factored the name out of `process_quiz_slot_tags`.