MDL-84987 qbank_managecategories: In-place editing of category names
Implement an inplace_editable widget for category names.

Currently, category names in the Edit categories interface are a link to
the questions in that category. This reduces that link so it only
applies to the question count, and replaces the static category name
with an inplace_editable widget.

As this is within a reactive component, and the category name is used
elsewhere for things like building the "Move category" modal, once the
inplace_editable returns a successful update we also update the category
name in the state.

The category component now listens for updates to the name in the state,
and updates the `data-categoryname` attribute in the DOM element. As
well as being triggered by the new inplace_editable, this is also
triggered when the name is edited in the modal editing form.